AI_DataChallenge/requirements.txt
# Core Data Science
pandas==1.5.3
numpy==1.24.3
scipy==1.10.1

# Machine Learning
scikit-learn==1.2.2
xgboost==1.7.5
optuna==3.1.1

# Geospatial Libraries
rasterio==1.3.6
GDAL==3.6.2
geopandas==0.12.2
shapely==2.0.1
pyproj==3.5.0
rioxarray==0.14.1

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1

# Jupyter Notebooks
jupyter==1.0.0
notebook==6.5.4
ipykernel==6.22.0
ipywidgets==8.0.6

# Data Warehouse Integration
snowflake-connector-python==3.0.2
snowflake-sqlalchemy==1.4.7

# Cloud and Remote Sensing
earthengine-api==0.1.350

# MLOps and Experiment Tracking
mlflow==2.3.0
wandb==0.15.0

# Utilities
tqdm==4.65.0
python-dotenv==1.0.0
pyyaml==6.0

# Optional: Additional Geospatial Tools
# fiona==1.9.3
# rtree==1.0.1
# geopy==2.3.0

# Optional: Advanced ML
# lightgbm==3.3.5
# catboost==1.2
# tensorflow==2.12.0
# torch==2.0.0
